+++ src/replay/index.ts
@@ -1,12 +1,4 @@
-export type { ReplayMode } from '#replay/replay-mode';
+// Types
 export type { Action, BaseAction, CaptureAction, MoveAction, PromoteAction } from '#types/actions';
-export type { BoardCoord, Square } from '#board/board-coords';
-export {
-    algebraicToCoords,
-    coordsToAlgebraic,
-    coordsToSquare,
-    coordsToSquareFromCoord,
-    squareToCoords,
-} from '#board/board-coords';
-export type { PlayerColor, PromotionToken } from '#types/tokens';
-export { isPromotionToken } from '#types/tokens';
+export type { Square } from '#board/board-coords';
+export type { PlayerColor } from '#types/tokens';